requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlmodel==0.0.14
pyyaml==6.0.1
numpy==1.24.3
pandas==2.0.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
httpx==0.25.2
